requirements.txt
crcmod==1.7
loguru==0.7.2
numpy==1.25.2
pandas==2.2.2
PyDbLite==3.0.4
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
PyYAML==6.0.1